+++ pulsar/apps/test/plugins/profile.py
@@ -129,7 +129,7 @@ class Profile(test.TestPlugin):
             local.tmp = tempfile.mktemp(dir=self.profile_temp_path)
             local.prof.enable()
         
-    def after_test_function_run(self, test, local, result, async):
+    def after_test_function_run(self, test, local, result):
         if self.config.profile:
             local.prof.disable()
             local.prof.dump_stats(local.tmp)